Extract interactive surface logic to resolver
Add a new InteractiveSurfaceResolver (InteractiveSurfaceResolver.cs) that encapsulates host detection, host-family/mode resolution, and interactive UI/TUI payload generation (including Spectre.Console menu handling). Remove the duplicated token arrays and host-detection helpers from SarifTools and replace the inline interactive-surface logic with a call to InteractiveSurfaceResolver.ResolveInteractiveSurface. Also update the SarifPrompts review description wording to better describe its purpose.
