+++ Sarifintown.AgentEngine/InteractiveSurfaceResolver.cs
@@ -1,0 +1,320 @@
+using ModelContextProtocol.Server;
+using System.Reflection;
+using System.Text.Json;
+
+namespace Sarifintown.AgentEngine
+{
+    internal static class InteractiveSurfaceResolver
+    {
+        private static readonly string[] IdeHostTokens =
+        [
+            "vscode",
+            "visualstudiocode",
+            "visualstudio",
+            "cursor",
+            "windsurf",
+            "jetbrains",
+            "rider",
+            "zed",
+            "eclipse",
+            "intellij",
+            "xcode"
+        ];
+
+        private static readonly string[] CliHostTokens =
+        [
+            "claudecode",
+            "claude",
+            "codex",
+            "aider",
+            "geminicli",
+            "opencode",
+            "terminal",
+            "bash",
+            "zsh",
+            "fish",
+            "powershell",
+            "pwsh",
+            "cmd",
+            "windowsterminal",
+            "iterm",
+            "tmux",
+            "kitty",
+            "alacritty"
+        ];
+
+        private static readonly string[] VsCodeFamilyTokens =
+        [
+            "vscode",
+            "visualstudiocode",
+            "cursor",
+            "windsurf"
+        ];
+
+        private static readonly string[] JetBrainsFamilyTokens =
+        [
+            "jetbrains",
+            "rider",
+            "intellij",
+            "pycharm",
+            "webstorm",
+            "clion",
+            "goland",
+            "rubymine"
+        ];
+
+        /// <summary>
+        /// Resolves host capabilities and returns the UI/TUI launch payload for the caller.
+        /// </summary>
+        public static string ResolveInteractiveSurface(
+            McpServer thisServer,
+            string hostHint,
+            bool startCliMenu,
+            Func<object> localHttpUiPayloadFactory)
+        {
+            ArgumentNullException.ThrowIfNull(localHttpUiPayloadFactory);
+
+            var host = DetectHost(thisServer, hostHint);
+            var mode = ResolveHostMode(host);
+            var hostFamily = ResolveHostFamily(host);
+            var usedFallback = string.Equals(host, "unknown", StringComparison.OrdinalIgnoreCase);
+
+            if (mode == "ide-ui")
+            {
+                return JsonSerializer.Serialize(new
+                {
+                    host,
+                    mode,
+                    host_family = hostFamily,
+                    uri = "ui://sarifintown/mcp/dashboard",
+                    local_http_ui = localHttpUiPayloadFactory(),
+                    bridge = new
+                    {
+                        transport = "postMessage",
+                        channel = "sarifintown.mcp.v1"
+                    },
+                    fallback = new
+                    {
+                        mode = "cli-tui",
+                        library = "Spectre.Console"
+                    }
+                });
+            }
+
+            string selectedAction = string.Empty;
+            string commandResult = string.Empty;
+            if (startCliMenu)
+            {
+                selectedAction = SpectreCliMenu.Start();
+                if (selectedAction.StartsWith("Triage ", StringComparison.Ordinal))
+                {
+                    commandResult = SpectreCliMenu
+                        .ExecuteTriageActionAsync(selectedAction)
+                        .GetAwaiter()
+                        .GetResult();
+                }
+            }
+
+            return JsonSerializer.Serialize(new
+            {
+                host,
+                mode,
+                host_family = hostFamily,
+                fallback_used = usedFallback,
+                local_http_ui = localHttpUiPayloadFactory(),
+                tui = new
+                {
+                    library = "Spectre.Console",
+                    action = selectedAction,
+                    action_result = commandResult,
+                    menu = "interactive"
+                }
+            });
+        }
+
+        private static string DetectHost(McpServer thisServer, string hostHint)
+        {
+            if (!string.IsNullOrWhiteSpace(hostHint))
+            {
+                return hostHint.Trim();
+            }
+
+            if (thisServer != null)
+            {
+                var hostFromServer = TryGetHostNameFromServer(thisServer);
+                if (!string.IsNullOrWhiteSpace(hostFromServer))
+                {
+                    return hostFromServer;
+                }
+            }
+
+            var hostFromEnvironment = TryGetHostNameFromEnvironment();
+            if (!string.IsNullOrWhiteSpace(hostFromEnvironment))
+            {
+                return hostFromEnvironment;
+            }
+
+            return "unknown";
+        }
+
+        private static string ResolveHostMode(string host)
+        {
+            if (string.IsNullOrWhiteSpace(host))
+            {
+                return "cli-tui";
+            }
+
+            var normalizedHost = NormalizeHost(host);
+
+            if (ContainsAnyToken(normalizedHost, IdeHostTokens))
+            {
+                return "ide-ui";
+            }
+
+            if (ContainsAnyToken(normalizedHost, CliHostTokens))
+            {
+                return "cli-tui";
+            }
+
+            return "cli-tui";
+        }
+
+        private static string ResolveHostFamily(string host)
+        {
+            if (string.IsNullOrWhiteSpace(host))
+            {
+                return "terminal-family";
+            }
+
+            var normalizedHost = NormalizeHost(host);
+
+            if (ContainsAnyToken(normalizedHost, VsCodeFamilyTokens))
+            {
+                return "vscode-family";
+            }
+
+            if (ContainsAnyToken(normalizedHost, JetBrainsFamilyTokens))
+            {
+                return "jetbrains-family";
+            }
+
+            if (normalizedHost.Contains("visualstudio", StringComparison.Ordinal))
+            {
+                return "visualstudio-family";
+            }
+
+            if (normalizedHost.Contains("zed", StringComparison.Ordinal))
+            {
+                return "zed-family";
+            }
+
+            if (ContainsAnyToken(normalizedHost, CliHostTokens))
+            {
+                return "terminal-family";
+            }
+
+            return "unknown-family";
+        }
+
+        private static string TryGetHostNameFromEnvironment()
+        {
+            var directValueVariables = new[]
+            {
+                "MCP_CLIENT_NAME",
+                "MCP_HOST",
+                "MCP_CLIENT",
+                "TERM_PROGRAM",
+                "TERM",
+                "TERMINAL_EMULATOR",
+                "PROMPT_TOOLKIT_SHELL",
+                "VSCODE_CWD",
+                "ELECTRON_RUN_AS_NODE"
+            };
+
+            foreach (var variable in directValueVariables)
+            {
+                var value = Environment.GetEnvironmentVariable(variable);
+                if (!string.IsNullOrWhiteSpace(value))
+                {
+                    return value.Trim();
+                }
+            }
+
+            if (!string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("CLAUDECODE")))
+            {
+                return "Claude Code";
+            }
+
+            if (!string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("CURSOR_TRACE_ID")))
+            {
+                return "Cursor";
+            }
+
+            if (!string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("VSCODE_GIT_IPC_HANDLE")))
+            {
+                return "Visual Studio Code";
+            }
+
+            return string.Empty;
+        }
+
+        private static string NormalizeHost(string host)
+        {
+            var lowered = host.Trim().ToLowerInvariant();
+
+            return string.Concat(lowered.Where(char.IsLetterOrDigit));
+        }
+
+        private static bool ContainsAnyToken(string normalizedHost, IEnumerable<string> tokens)
+        {
+            return tokens.Any(token => normalizedHost.Contains(token, StringComparison.Ordinal));
+        }
+
+        private static string TryGetHostNameFromServer(McpServer server)
+        {
+            var directClientInfo = GetPropertyValue(server, "ClientInfo");
+            var directName = GetPropertyValue(directClientInfo, "Name")?.ToString();
+
+            if (!string.IsNullOrWhiteSpace(directName))
+            {
+                return directName.Trim();
+            }
+
+            var initializeRequest = GetPropertyValue(server, "InitializeRequest");
+            var requestClientInfo = GetPropertyValue(initializeRequest, "ClientInfo");
+            var requestClientName = GetPropertyValue(requestClientInfo, "Name")?.ToString();
+
+            if (!string.IsNullOrWhiteSpace(requestClientName))
+            {
+                return requestClientName.Trim();
+            }
+
+            var session = GetPropertyValue(server, "Session");
+            var sessionClientInfo = GetPropertyValue(session, "ClientInfo");
+            var sessionClientName = GetPropertyValue(sessionClientInfo, "Name")?.ToString();
+
+            return string.IsNullOrWhiteSpace(sessionClientName)
+                ? string.Empty
+                : sessionClientName.Trim();
+        }
+
+        private static object? GetPropertyValue(object? instance, string propertyName)
+        {
+            if (instance == null)
+            {
+                return null;
+            }
+
+            var property = instance
+                .GetType()
+                .GetProperty(propertyName, BindingFlags.Instance | BindingFlags.Public | BindingFlags.IgnoreCase);
+
+            if (property == null)
+            {
+                return null;
+            }
+
+            return property.GetValue(instance);
+        }
+    }
+}
